requirements.txt
Python3.7
mysql-connector==2.2.9
requests==2.22.0